Refactors charm for adding multiple relations support
Updates IngressBrokenEvent to be based on RelationEvent, and
include the relation in the event data, which will help us
determine which relation has been broken.

Separates the config / relation data into the _ConfigOrRelation object,
which receives the config and a relation in its constructor. This will allow
us to get necessary data from a specific relation (which is overriden by the
config options).