Adds ExternalName Service type option
Kubernetes has a Service type `ExternalName`, which allows the Cluster DNS
Service to return a CNAME record to the set `externalName`. This can be useful
in certain situations.

The `nginx-ingress-integrator` charm does not have any support for this at
this moment.

This commit adds the `external-name` relation field and config option, which,
if set, will cause the `nginx-ingress-integrator` charm to create a Kubernetes
ExternalName Service with the `externalName` set instead of a ClusterIP Service.